+++ vector/src/main/java/im/vector/VectorApp.java
@@ -194,6 +194,11 @@ public class VectorApp extends MultiDexApplication {
         Log.d(LOG_TAG, "onCreate");
         super.onCreate();
 
+        PreferencesManager.setIntegrationManagerUrls(this,
+                getString(R.string.integrations_ui_url),
+                getString(R.string.integrations_rest_url),
+                getString(R.string.integrations_jitsi_widget_url));
+
         mLifeCycleListener = new VectorLifeCycleObserver();
         ProcessLifecycleOwner.get().getLifecycle().addObserver(mLifeCycleListener);