+++ vector/src/main/java/im/vector/widgets/WidgetsManager.java
@@ -19,7 +19,6 @@
 package im.vector.widgets;
 
 import android.content.Context;
-import android.support.v7.preference.PreferenceManager;
 import android.text.TextUtils;
 
 import com.google.gson.JsonObject;
@@ -30,8 +29,10 @@ import org.matrix.androidsdk.core.callback.ApiCallback;
 import org.matrix.androidsdk.core.callback.SimpleApiCallback;
 import org.matrix.androidsdk.core.model.MatrixError;
 import org.matrix.androidsdk.data.Room;
+import org.matrix.androidsdk.features.terms.TermsNotSignedException;
 import org.matrix.androidsdk.rest.model.Event;
 
+import java.net.HttpURLConnection;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -48,6 +49,7 @@ import im.vector.R;
 import im.vector.VectorApp;
 import im.vector.extensions.UrlExtensionsKt;
 import im.vector.settings.VectorLocale;
+import im.vector.widgets.tokens.TokensStore;
 
 public class WidgetsManager {
     private static final String LOG_TAG = WidgetsManager.class.getSimpleName();
@@ -67,11 +69,6 @@ public class WidgetsManager {
      */
     private static final String WIDGET_TYPE_JITSI = "jitsi";
 
-    /**
-     * Widget preferences
-     */
-    private static final String SCALAR_TOKEN_PREFERENCE_KEY = "SCALAR_TOKEN_PREFERENCE_KEY";
-
     private final IntegrationManagerConfig config;
 
     public WidgetsManager(IntegrationManagerConfig config) {
@@ -81,6 +78,7 @@ public class WidgetsManager {
     public String getUIUrl() {
         return config.getUiUrl();
     }
+
     /**
      * Widget error code
      */
@@ -530,12 +528,37 @@ public class WidgetsManager {
      * @param callback the asynchronous callback
      */
     public void getScalarToken(final Context context, final MXSession session, final ApiCallback<String> callback) {
-        final String preferenceKey = SCALAR_TOKEN_PREFERENCE_KEY + session.getMyUserId();
+        final TokensStore tokensStore = new TokensStore(context);
 
-        final String scalarToken = PreferenceManager.getDefaultSharedPreferences(context).getString(preferenceKey, null);
+        final String scalarToken = tokensStore.getToken(session.getMyUserId(), config.getApiUrl());
 
         if (null != scalarToken) {
-            callback.onSuccess(scalarToken);
+            WidgetsRestClient widgetsRestClient = new WidgetsRestClient(context, config);
+            widgetsRestClient.validateToken(scalarToken, new SimpleApiCallback<Map<String, String>>(callback) {
+
+                @Override
+                public void onSuccess(Map<String, String> info) {
+                    if (null != callback) {
+                        callback.onSuccess(scalarToken);
+                    }
+                }
+
+                @Override
+                public void onMatrixError(MatrixError e) {
+                    if (MatrixError.TERMS_NOT_SIGNED.equals(e.errcode)) {
+                        if (null != callback) {
+                            callback.onUnexpectedError(new TermsNotSignedException(scalarToken));
+                        }
+                    } else if (e.mStatus == HttpURLConnection.HTTP_FORBIDDEN /* 403 */) {
+                        // Refresh the token
+                        Log.w(LOG_TAG, "Invalid token, clear it and get a new token");
+                        clearScalarToken(context, session);
+                        getScalarToken(context, session, callback);
+                    } else {
+                        super.onMatrixError(e);
+                    }
+                }
+            });
         } else {
             session.openIdToken(new SimpleApiCallback<Map<Object, Object>>(callback) {
                 @Override
@@ -548,15 +571,29 @@ public class WidgetsManager {
                             String token = response.get("scalar_token");
 
                             if (null != token) {
-                                PreferenceManager.getDefaultSharedPreferences(context)
-                                        .edit()
-                                        .putString(preferenceKey, token)
-                                        .apply();
+                                tokensStore.setToken(session.getMyUserId(), config.getApiUrl(), token);
                             }
 
-                            if (null != callback) {
-                                callback.onSuccess(token);
-                            }
+                            // Validate it (this mostly checks to see if the IM needs us to agree to some terms)
+
+                            widgetsRestClient.validateToken(token, new SimpleApiCallback<Map<String, String>>(callback) {
+                                @Override
+                                public void onSuccess(Map<String, String> info) {
+                                    if (null != callback) {
+                                        callback.onSuccess(token);
+                                    }
+                                }
+
+                                @Override
+                                public void onMatrixError(MatrixError e) {
+                                    if (MatrixError.TERMS_NOT_SIGNED.equals(e.errcode)) {
+                                        callback.onUnexpectedError(new TermsNotSignedException(token));
+                                    } else {
+                                        super.onMatrixError(e);
+                                    }
+                                }
+                            });
+
                         }
                     });
                 }
@@ -571,11 +608,6 @@ public class WidgetsManager {
      * @param session current session, to retrieve the current user
      */
     public void clearScalarToken(Context context, final MXSession session) {
-        final String preferenceKey = SCALAR_TOKEN_PREFERENCE_KEY + session.getMyUserId();
-
-        PreferenceManager.getDefaultSharedPreferences(context)
-                .edit()
-                .remove(preferenceKey)
-                .apply();
+        new TokensStore(context).clear();
     }
 }